Rename component and clean up
